requirements.txt
kivy
kivymd
requests
plyer
openai
python-dotenv
nvidia-riva-client
